backend/requirements.txt
pylint==2.1.1
pylint-django==2.0.2
flake8==3.6.0
django==2.0.9
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
